rename 'maintainInvariant.js' to 'maintainState.js'
